clean(realized I didn't need to render passes for my triangle shenanigans): This is kinda silly, but I used two render passes because in my head. Render Pass == Layer.
Since I wanted the background to be displayed on the bottom layer,
and the triangle on the second layer it made sense to use two render passes (layers).

But since I clear and then draw a shader on top of it, I was already getting that visual effect
.

Had that realization at 2:10am (as I was thinking about the code I had written and reviewing the ideas) and realized (wait a minute, am I dumb?)

Turns out that yes... Anyhow, simplified the code by removing the unneeded render passes.